replaced some synchronized calls with concurrent collections or ReentrantLock
