+++ cpo-core/src/main/java/org/synchronoss/cpo/CpoException.java
@@ -30,37 +30,28 @@ package org.synchronoss.cpo;
  */
 public class CpoException extends Exception {
 
-  /** Version Id for this class. */
-  private static final long serialVersionUID = 1L;
-
-  /**
-   * Nested Exception to hold wrapped exception.
-   *
-   * @serial
-   */
-  private Throwable detail;
-
   /** Constructs a <code>CpoException</code> with no specified detail message. */
-  public CpoException() {}
+  public CpoException() {
+    super();
+  }
 
   /**
    * Constructs a <code>CpoException</code> with the specified detail message.
    *
-   * @param s the detail message
+   * @param message the detail message
    */
-  public CpoException(String s) {
-    super(s);
+  public CpoException(String message) {
+    super(message);
   }
 
   /**
    * Constructs a <code>CpoException</code> with the specified detail message and nested exception.
    *
-   * @param s the detail message
-   * @param ex the nested exception
+   * @param message the detail message
+   * @param cause the nested exception
    */
-  public CpoException(String s, Throwable ex) {
-    super(s);
-    detail = ex;
+  public CpoException(String message, Throwable cause) {
+    super(message, cause);
   }
 
   /**
@@ -69,83 +60,6 @@ public class CpoException extends Exception {
    * @param ex the nested exception
    */
   public CpoException(Throwable ex) {
-    super();
-    detail = ex;
-  }
-
-  /**
-   * Returns the detail message, including the message from the nested exception if there is one.
-   */
-  @Override
-  public String getMessage() {
-    StringBuilder msg = new StringBuilder("\n");
-
-    msg.append(super.getMessage());
-
-    if (detail != null) {
-      msg.append("\n");
-      msg.append(detail.getMessage());
-      if (detail.getCause() != null) {
-        msg.append(detail.getCause().getMessage());
-      }
-    }
-    return msg.toString();
-  }
-
-  /**
-   * Returns the detail message, including the message from the nested exception if there is one.
-   */
-  @Override
-  public String getLocalizedMessage() {
-    StringBuilder msg = new StringBuilder("\n");
-
-    msg.append(super.getLocalizedMessage());
-
-    if (detail != null) {
-      msg.append("\n");
-      msg.append(detail.getLocalizedMessage());
-      if (detail.getCause() != null) {
-        msg.append(detail.getCause().getLocalizedMessage());
-      }
-    }
-    return msg.toString();
-  }
-
-  /**
-   * Prints the composite message and the embedded stack trace to the specified stream <code>ps
-   * </code>.
-   *
-   * @param ps the print stream
-   */
-  @Override
-  public void printStackTrace(java.io.PrintStream ps) {
-    synchronized (ps) {
-      if (detail != null) {
-        detail.printStackTrace(ps);
-      }
-      super.printStackTrace(ps);
-    }
-  }
-
-  /** Prints the composite message to <code>System.err</code>. */
-  @Override
-  public void printStackTrace() {
-    printStackTrace(System.err);
-  }
-
-  /**
-   * Prints the composite message and the embedded stack trace to the specified print writer <code>
-   * pw</code>.
-   *
-   * @param pw the print writer
-   */
-  @Override
-  public void printStackTrace(java.io.PrintWriter pw) {
-    synchronized (pw) {
-      if (detail != null) {
-        detail.printStackTrace(pw);
-      }
-      super.printStackTrace(pw);
-    }
+    super(ex);
   }
 }
