+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/adapter/BlobTest.java
@@ -1,4 +1,4 @@
-package org.synchronoss.cpo.jdbc;
+package org.synchronoss.cpo.jdbc.adapter;
 
 /*-
  * [[
@@ -28,7 +28,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.synchronoss.cpo.CpoAdapter;
 import org.synchronoss.cpo.CpoAdapterFactoryManager;
-import org.synchronoss.cpo.CpoTrxAdapter;
+import org.synchronoss.cpo.jdbc.LobValueObject;
 import org.synchronoss.cpo.jdbc.meta.JdbcCpoMetaDescriptor;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -40,27 +40,25 @@ import org.testng.annotations.Test;
  *
  * @author david berry
  */
-public class BlobTrxTest {
+public class BlobTest {
 
-  private static final Logger logger = LoggerFactory.getLogger(BlobTrxTest.class);
-  private static int BLOB_SIZE = 64999;
-  private CpoAdapter cpoAdapter = null;
-  private CpoTrxAdapter trxAdapter = null;
+  private static final Logger logger = LoggerFactory.getLogger(BlobTest.class);
   private JdbcCpoMetaDescriptor metaDescriptor = null;
+  private CpoAdapter cpoAdapter = null;
   private byte[] testBlob = null;
   private char[] testClob = "This is a test Clob used for testing clobs".toCharArray();
   private byte[] testBlob2 = null;
   private char[] testClob2 = "This is a second test Clob used for testing clobs".toCharArray();
   private boolean isSupportsBlobs = true;
 
-  public BlobTrxTest() {}
+  public BlobTest() {}
 
   /**
    * <code>setUp</code> Load the datasource from the properties in the property file
    * jdbc_en_US.properties
    *
    * @author david berry
-   * @version '$Id: BlobTestTrx.java,v 1.2 2006/01/31 22:31:06 dberry Exp $'
+   * @version '$Id: BlobTest.java,v 1.15 2006/02/15 18:34:19 dberry Exp $'
    */
   @Parameters({"db.blobsupport"})
   @BeforeClass
@@ -70,8 +68,7 @@ public class BlobTrxTest {
 
     try {
       cpoAdapter = CpoAdapterFactoryManager.getCpoAdapter(JdbcStatics.ADAPTER_CONTEXT_JDBC);
-      trxAdapter = CpoAdapterFactoryManager.getCpoTrxAdapter(JdbcStatics.ADAPTER_CONTEXT_JDBC);
-      assertNotNull(trxAdapter, method + "trxAdapter is null");
+      assertNotNull(cpoAdapter, method + "cpoAdapter is null");
       metaDescriptor = (JdbcCpoMetaDescriptor) cpoAdapter.getCpoMetaDescriptor();
     } catch (Exception e) {
       fail(method + e.getMessage());
@@ -79,17 +76,17 @@ public class BlobTrxTest {
   }
 
   @Test
-  public void testGZipBlobInsertandDeleteTrx() {
+  public void testTrxGZipBlobInsertandDelete() {
 
     if (isSupportsBlobs) {
 
-      testBlob = new byte[BLOB_SIZE];
-      for (int i = 0; i < BLOB_SIZE; i++) {
+      testBlob = new byte[JdbcStatics.BLOB_SIZE];
+      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
         testBlob[i] = (byte) (((int) 'a') + (i % 26));
       }
 
-      testBlob2 = new byte[BLOB_SIZE];
-      for (int i = 0; i < BLOB_SIZE; i++) {
+      testBlob2 = new byte[JdbcStatics.BLOB_SIZE];
+      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
         testBlob2[i] = (byte) (((int) 'z') - (i % 26));
       }
 
@@ -99,35 +96,30 @@ public class BlobTrxTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
       } catch (Exception ie) {
         logger.error("error deleting lob");
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(lvo.getBLob());
         String blob2 = new String(lvo2.getBLob());
 
         assertEquals(blob1, blob2);
+
+        String clob1 = new String(lvo.getCLob());
+        String clob2 = new String(lvo2.getCLob());
+
+        assertEquals(clob1, clob2);
 
       } catch (Exception ie) {
         logger.error("error retrieving lob", ie);
@@ -137,21 +129,20 @@ public class BlobTrxTest {
       try {
         lvo2.setBLob(testBlob2);
         lvo2.setCLob(testClob2);
-        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        trxAdapter.commit();
-
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(testBlob2);
         String blob2 = new String(lvo2.getBLob());
 
         assertEquals(blob1, blob2);
 
+        String clob1 = new String(testClob2);
+        String clob2 = new String(lvo2.getCLob());
+
+        assertEquals(clob1, clob2);
+
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
     } else {
@@ -160,17 +151,17 @@ public class BlobTrxTest {
   }
 
   @Test
-  public void testBlobInsertandDeleteTrx() {
+  public void testTrxBlobInsertandDelete() {
 
     if (isSupportsBlobs) {
 
-      testBlob = new byte[BLOB_SIZE];
-      for (int i = 0; i < BLOB_SIZE; i++) {
+      testBlob = new byte[JdbcStatics.BLOB_SIZE];
+      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
         testBlob[i] = (byte) (((int) 'a') + (i % 26));
       }
 
-      testBlob2 = new byte[BLOB_SIZE];
-      for (int i = 0; i < BLOB_SIZE; i++) {
+      testBlob2 = new byte[JdbcStatics.BLOB_SIZE];
+      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
         testBlob2[i] = (byte) (((int) 'z') - (i % 26));
       }
 
@@ -180,31 +171,21 @@ public class BlobTrxTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
       } catch (Exception ie) {
         logger.error("error deleting lob");
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(lvo.getBLob2());
         String blob2 = new String(lvo2.getBLob2());
 
@@ -217,9 +198,8 @@ public class BlobTrxTest {
 
       try {
         lvo2.setBLob2(testBlob);
-        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        trxAdapter.commit();
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(testBlob);
         String blob2 = new String(lvo2.getBLob2());
 
@@ -227,10 +207,6 @@ public class BlobTrxTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
     } else {
@@ -262,7 +238,7 @@ public class BlobTrxTest {
    * }
    */
   @Test
-  public void testEmptyGZipBlobInsertandDeleteTrx() {
+  public void testTrxEmptyGZipBlobInsertandDelete() {
 
     if (isSupportsBlobs) {
 
@@ -275,31 +251,21 @@ public class BlobTrxTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
       } catch (Exception ie) {
         logger.error("error deleting lob");
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         byte blob1[] = lvo.getBLob();
         byte blob2[] = lvo2.getBLob();
 
@@ -316,9 +282,8 @@ public class BlobTrxTest {
       try {
         lvo2.setBLob(testBlob2);
         lvo2.setCLob(testClob2);
-        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        trxAdapter.commit();
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         byte blob1[] = testBlob2;
         byte blob2[] = lvo2.getBLob();
 
@@ -329,10 +294,6 @@ public class BlobTrxTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
     } else {
@@ -341,7 +302,7 @@ public class BlobTrxTest {
   }
 
   @Test
-  public void testNullGZipBlobInsertandDeleteTrx() {
+  public void testTrxNullGZipBlobInsertandDelete() {
 
     if (isSupportsBlobs) {
 
@@ -351,31 +312,21 @@ public class BlobTrxTest {
       lvo.setBLob2(null);
 
       try {
-        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
       } catch (Exception ie) {
         logger.error("error deleting lob");
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
-        trxAdapter.commit();
+        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
 
         assertNull(lvo2.getBLob());
         assertNull(lvo2.getBLob2());
@@ -389,9 +340,8 @@ public class BlobTrxTest {
       try {
         lvo2.setBLob(null);
         lvo2.setCLob(null);
-        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        trxAdapter.commit();
-        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
 
         assertNull(lvo2.getBLob());
         assertNull(lvo2.getBLob2());
@@ -399,10 +349,6 @@ public class BlobTrxTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
-        try {
-          trxAdapter.rollback();
-        } catch (Exception e) {
-        }
         fail(ie.getMessage());
       }
     } else {
@@ -412,11 +358,8 @@ public class BlobTrxTest {
 
   @AfterClass
   public void tearDown() {
-    try {
-      trxAdapter.close();
-    } catch (Exception e) {
-    }
-    trxAdapter = null;
+
+    cpoAdapter = null;
   }
 
   private boolean isEqual(byte[] b1, byte[] b2) {
