+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/adapter/BlobTrxTest.java
@@ -1,4 +1,4 @@
-package org.synchronoss.cpo.jdbc;
+package org.synchronoss.cpo.jdbc.adapter;
 
 /*-
  * [[
@@ -28,6 +28,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.synchronoss.cpo.CpoAdapter;
 import org.synchronoss.cpo.CpoAdapterFactoryManager;
+import org.synchronoss.cpo.CpoTrxAdapter;
+import org.synchronoss.cpo.jdbc.LobValueObject;
 import org.synchronoss.cpo.jdbc.meta.JdbcCpoMetaDescriptor;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -39,25 +41,27 @@ import org.testng.annotations.Test;
  *
  * @author david berry
  */
-public class BlobTest {
+public class BlobTrxTest {
 
-  private static final Logger logger = LoggerFactory.getLogger(BlobTest.class);
-  private JdbcCpoMetaDescriptor metaDescriptor = null;
+  private static final Logger logger = LoggerFactory.getLogger(BlobTrxTest.class);
+  private static int BLOB_SIZE = 64999;
   private CpoAdapter cpoAdapter = null;
+  private CpoTrxAdapter trxAdapter = null;
+  private JdbcCpoMetaDescriptor metaDescriptor = null;
   private byte[] testBlob = null;
   private char[] testClob = "This is a test Clob used for testing clobs".toCharArray();
   private byte[] testBlob2 = null;
   private char[] testClob2 = "This is a second test Clob used for testing clobs".toCharArray();
   private boolean isSupportsBlobs = true;
 
-  public BlobTest() {}
+  public BlobTrxTest() {}
 
   /**
    * <code>setUp</code> Load the datasource from the properties in the property file
    * jdbc_en_US.properties
    *
    * @author david berry
-   * @version '$Id: BlobTest.java,v 1.15 2006/02/15 18:34:19 dberry Exp $'
+   * @version '$Id: BlobTestTrx.java,v 1.2 2006/01/31 22:31:06 dberry Exp $'
    */
   @Parameters({"db.blobsupport"})
   @BeforeClass
@@ -67,7 +71,8 @@ public class BlobTest {
 
     try {
       cpoAdapter = CpoAdapterFactoryManager.getCpoAdapter(JdbcStatics.ADAPTER_CONTEXT_JDBC);
-      assertNotNull(cpoAdapter, method + "cpoAdapter is null");
+      trxAdapter = CpoAdapterFactoryManager.getCpoTrxAdapter(JdbcStatics.ADAPTER_CONTEXT_JDBC);
+      assertNotNull(trxAdapter, method + "trxAdapter is null");
       metaDescriptor = (JdbcCpoMetaDescriptor) cpoAdapter.getCpoMetaDescriptor();
     } catch (Exception e) {
       fail(method + e.getMessage());
@@ -75,17 +80,17 @@ public class BlobTest {
   }
 
   @Test
-  public void testTrxGZipBlobInsertandDelete() {
+  public void testGZipBlobInsertandDeleteTrx() {
 
     if (isSupportsBlobs) {
 
-      testBlob = new byte[JdbcStatics.BLOB_SIZE];
-      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
+      testBlob = new byte[BLOB_SIZE];
+      for (int i = 0; i < BLOB_SIZE; i++) {
         testBlob[i] = (byte) (((int) 'a') + (i % 26));
       }
 
-      testBlob2 = new byte[JdbcStatics.BLOB_SIZE];
-      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
+      testBlob2 = new byte[BLOB_SIZE];
+      for (int i = 0; i < BLOB_SIZE; i++) {
         testBlob2[i] = (byte) (((int) 'z') - (i % 26));
       }
 
@@ -95,30 +100,35 @@ public class BlobTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error deleting lob");
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(lvo.getBLob());
         String blob2 = new String(lvo2.getBLob());
 
         assertEquals(blob1, blob2);
-
-        String clob1 = new String(lvo.getCLob());
-        String clob2 = new String(lvo2.getCLob());
-
-        assertEquals(clob1, clob2);
 
       } catch (Exception ie) {
         logger.error("error retrieving lob", ie);
@@ -128,20 +138,21 @@ public class BlobTest {
       try {
         lvo2.setBLob(testBlob2);
         lvo2.setCLob(testClob2);
-        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        trxAdapter.commit();
+
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(testBlob2);
         String blob2 = new String(lvo2.getBLob());
 
         assertEquals(blob1, blob2);
 
-        String clob1 = new String(testClob2);
-        String clob2 = new String(lvo2.getCLob());
-
-        assertEquals(clob1, clob2);
-
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
     } else {
@@ -150,17 +161,17 @@ public class BlobTest {
   }
 
   @Test
-  public void testTrxBlobInsertandDelete() {
+  public void testBlobInsertandDeleteTrx() {
 
     if (isSupportsBlobs) {
 
-      testBlob = new byte[JdbcStatics.BLOB_SIZE];
-      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
+      testBlob = new byte[BLOB_SIZE];
+      for (int i = 0; i < BLOB_SIZE; i++) {
         testBlob[i] = (byte) (((int) 'a') + (i % 26));
       }
 
-      testBlob2 = new byte[JdbcStatics.BLOB_SIZE];
-      for (int i = 0; i < JdbcStatics.BLOB_SIZE; i++) {
+      testBlob2 = new byte[BLOB_SIZE];
+      for (int i = 0; i < BLOB_SIZE; i++) {
         testBlob2[i] = (byte) (((int) 'z') - (i % 26));
       }
 
@@ -170,21 +181,31 @@ public class BlobTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error deleting lob");
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(lvo.getBLob2());
         String blob2 = new String(lvo2.getBLob2());
 
@@ -197,8 +218,9 @@ public class BlobTest {
 
       try {
         lvo2.setBLob2(testBlob);
-        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        trxAdapter.commit();
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         String blob1 = new String(testBlob);
         String blob2 = new String(lvo2.getBLob2());
 
@@ -206,6 +228,10 @@ public class BlobTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
     } else {
@@ -237,7 +263,7 @@ public class BlobTest {
    * }
    */
   @Test
-  public void testTrxEmptyGZipBlobInsertandDelete() {
+  public void testEmptyGZipBlobInsertandDeleteTrx() {
 
     if (isSupportsBlobs) {
 
@@ -250,21 +276,31 @@ public class BlobTest {
       lvo.setBLob2(testBlob2);
 
       try {
-        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error deleting lob");
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         byte blob1[] = lvo.getBLob();
         byte blob2[] = lvo2.getBLob();
 
@@ -281,8 +317,9 @@ public class BlobTest {
       try {
         lvo2.setBLob(testBlob2);
         lvo2.setCLob(testClob2);
-        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        trxAdapter.commit();
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
         byte blob1[] = testBlob2;
         byte blob2[] = lvo2.getBLob();
 
@@ -293,6 +330,10 @@ public class BlobTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
     } else {
@@ -301,7 +342,7 @@ public class BlobTest {
   }
 
   @Test
-  public void testTrxNullGZipBlobInsertandDelete() {
+  public void testNullGZipBlobInsertandDeleteTrx() {
 
     if (isSupportsBlobs) {
 
@@ -311,21 +352,31 @@ public class BlobTest {
       lvo.setBLob2(null);
 
       try {
-        cpoAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.deleteBean(LobValueObject.FG_DELETE_DELETELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error deleting lob");
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        cpoAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.insertBean(LobValueObject.FG_CREATE_CREATELVO, lvo);
+        trxAdapter.commit();
       } catch (Exception ie) {
         logger.error("error inserting lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
 
       try {
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
 
         assertNull(lvo2.getBLob());
         assertNull(lvo2.getBLob2());
@@ -339,8 +390,9 @@ public class BlobTest {
       try {
         lvo2.setBLob(null);
         lvo2.setCLob(null);
-        cpoAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
-        lvo2 = cpoAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
+        trxAdapter.updateBean(LobValueObject.FG_UPDATE_UPDATELVO, lvo2);
+        trxAdapter.commit();
+        lvo2 = trxAdapter.retrieveBean(LobValueObject.FG_RETRIEVE_RETRIEVELVO, lvo);
 
         assertNull(lvo2.getBLob());
         assertNull(lvo2.getBLob2());
@@ -348,6 +400,10 @@ public class BlobTest {
 
       } catch (Exception ie) {
         logger.error("error updating lob", ie);
+        try {
+          trxAdapter.rollback();
+        } catch (Exception e) {
+        }
         fail(ie.getMessage());
       }
     } else {
@@ -357,8 +413,11 @@ public class BlobTest {
 
   @AfterClass
   public void tearDown() {
-
-    cpoAdapter = null;
+    try {
+      trxAdapter.close();
+    } catch (Exception e) {
+    }
+    trxAdapter = null;
   }
 
   private boolean isEqual(byte[] b1, byte[] b2) {
