+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/adapter/DeleteObjectTest.java
@@ -1,4 +1,4 @@
-package org.synchronoss.cpo.jdbc;
+package org.synchronoss.cpo.jdbc.adapter;
 
 /*-
  * [[
@@ -24,7 +24,9 @@ package org.synchronoss.cpo.jdbc;
 
 import static org.testng.Assert.*;
 
+import java.sql.Timestamp;
 import java.util.ArrayList;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.synchronoss.cpo.CpoAdapter;
@@ -33,32 +35,38 @@ import org.synchronoss.cpo.CpoWhere;
 import org.synchronoss.cpo.enums.Comparison;
 import org.synchronoss.cpo.enums.Logical;
 import org.synchronoss.cpo.helper.ExceptionHelper;
+import org.synchronoss.cpo.jdbc.ValueObject;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
 /**
- * ExistObjectTest is a test class for the exists api calls
+ * DeleteObjectTest is a test class for testing the JdbcAdapter deleteObject method
  *
  * @author david berry
  */
-public class ExistObjectTest {
+public class DeleteObjectTest {
 
-  private static final Logger logger = LoggerFactory.getLogger(ExistObjectTest.class);
+  private static final Logger logger = LoggerFactory.getLogger(DeleteObjectTest.class);
+  private final ArrayList<ValueObject> al = new ArrayList<>();
   private CpoAdapter cpoAdapter = null;
+  private boolean isSupportsMillis = true;
 
-  public ExistObjectTest() {}
+  public DeleteObjectTest() {}
 
   /**
    * <code>setUp</code> Load the datasource from the properties in the property file
    * jdbc_en_US.properties
    *
    * @author david berry
-   * @version '$Id: ExistObjectTest.java,v 1.2 2006/01/30 19:09:23 dberry Exp $'
+   * @version '$Id: InsertObjectTest.java,v 1.3 2006/01/30 19:09:23 dberry Exp $'
    */
+  @Parameters({"db.millisupport"})
   @BeforeClass
-  public void setUp() {
+  public void setUp(boolean milliSupport) {
     String method = "setUp:";
+    isSupportsMillis = milliSupport;
 
     try {
       cpoAdapter = CpoAdapterFactoryManager.getCpoAdapter(JdbcStatics.ADAPTER_CONTEXT_JDBC);
@@ -66,76 +74,66 @@ public class ExistObjectTest {
     } catch (Exception e) {
       fail(method + e.getMessage());
     }
-    ValueObject vo = ValueObjectFactory.createValueObject(1);
-    vo.setAttrVarChar("WHERE");
+  }
+
+  @Test
+  public void testDeleteObject() {
+    String method = "testDeleteObject:";
+    ValueObject valObj = ValueObjectFactory.createValueObject(5);
+
+    valObj.setAttrVarChar("testDelete");
+    valObj.setAttrInteger(3);
+    Timestamp ts = new Timestamp(System.currentTimeMillis());
+
+    if (!isSupportsMillis) {
+      ts.setNanos(0);
+    }
+
+    valObj.setAttrDatetime(ts);
+
+    valObj.setAttrBit(true);
+
+    al.add(valObj);
 
     try {
-      long count = cpoAdapter.insertBean(vo);
-      assertEquals(1, count, "Should be inserted");
+      cpoAdapter.insertBean(valObj);
     } catch (Exception e) {
       logger.error(ExceptionHelper.getLocalizedMessage(e));
       fail(method + e.getMessage());
     }
-  }
 
-  @Test
-  public void testExistObject() {
-    String method = "testExistObject:";
-
+    // try the where on the delete, should delete 0
     try {
-      ValueObject valObj = ValueObjectFactory.createValueObject(1);
-      long count = cpoAdapter.existsBean(valObj);
-      assertEquals(1, count, "Object not Found");
+      List<CpoWhere> cws = new ArrayList<>();
+      cws.add(cpoAdapter.newWhere(Logical.AND, ValueObject.ATTR_ID, Comparison.EQ, 2));
+      long deleted = cpoAdapter.deleteBean(ValueObject.FG_DELETE_NULL, valObj, cws, null, null);
+      assertEquals(deleted, 0, "Should not have deleted anything");
     } catch (Exception e) {
+      logger.error(ExceptionHelper.getLocalizedMessage(e));
       fail(method + e.getMessage());
     }
 
+    // try the where on the delete, should delete 1
     try {
-      ValueObject valObj = ValueObjectFactory.createValueObject(5);
-      long count = cpoAdapter.existsBean(valObj);
-      assertEquals(0, count, "Object Found");
+      List<CpoWhere> cws = new ArrayList<>();
+      cws.add(cpoAdapter.newWhere(Logical.OR, ValueObject.ATTR_ID, Comparison.EQ, 2));
+      long deleted = cpoAdapter.deleteBean(ValueObject.FG_DELETE_NULL, valObj, cws, null, null);
+      assertEquals(deleted, 1, "Should have deleted 1");
     } catch (Exception e) {
-      fail(method + e.getMessage());
-    }
-  }
-
-  @Test
-  public void testExistObjectWhere() {
-    String method = "testExistObjectWhere:";
-
-    try {
-      ValueObject valObj = ValueObjectFactory.createValueObject(1);
-      CpoWhere where =
-          cpoAdapter.newWhere(Logical.AND, ValueObject.ATTR_ATTRVARCHAR, Comparison.EQ, "WHERE");
-      ArrayList<CpoWhere> wheres = new ArrayList<>();
-      wheres.add(where);
-      long count = cpoAdapter.existsBean(ValueObject.FG_EXIST_NULL, valObj, wheres);
-      assertEquals(1, count, "Object not Found");
-    } catch (Exception e) {
-      fail(method + e.getMessage());
-    }
-
-    try {
-      ValueObject valObj = ValueObjectFactory.createValueObject(1);
-      CpoWhere where =
-          cpoAdapter.newWhere(Logical.AND, ValueObject.ATTR_ATTRVARCHAR, Comparison.EQ, "NOWHERE");
-      ArrayList<CpoWhere> wheres = new ArrayList<>();
-      wheres.add(where);
-      long count = cpoAdapter.existsBean(ValueObject.FG_EXIST_NULL, valObj, wheres);
-      assertEquals(0, count, "Object Found");
-    } catch (Exception e) {
+      logger.error(ExceptionHelper.getLocalizedMessage(e));
       fail(method + e.getMessage());
     }
   }
 
   @AfterClass
   public void tearDown() {
-    ValueObject vo = ValueObjectFactory.createValueObject(1);
+    String method = "tearDown:";
     try {
-      long count = cpoAdapter.deleteBean(vo);
-      assertEquals(1, count, "Should be deleted");
+      cpoAdapter.deleteBeans(al);
+
     } catch (Exception e) {
       logger.error(ExceptionHelper.getLocalizedMessage(e));
+      fail(method + e.getMessage());
     }
     cpoAdapter = null;
   }
