+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/adapter/JdbcStatics.java
@@ -1,4 +1,4 @@
-package org.synchronoss.cpo.jdbc;
+package org.synchronoss.cpo.jdbc.adapter;
 
 /*-
  * [[
