+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/exporter/ExporterTest.java
@@ -42,9 +42,9 @@ import org.synchronoss.cpo.exporter.CpoClassSourceGenerator;
 import org.synchronoss.cpo.exporter.CpoInterfaceSourceGenerator;
 import org.synchronoss.cpo.exporter.CpoLegacyClassSourceGenerator;
 import org.synchronoss.cpo.exporter.MetaXmlObjectExporter;
-import org.synchronoss.cpo.jdbc.JdbcStatics;
 import org.synchronoss.cpo.jdbc.ValueObject;
-import org.synchronoss.cpo.jdbc.ValueObjectFactory;
+import org.synchronoss.cpo.jdbc.adapter.JdbcStatics;
+import org.synchronoss.cpo.jdbc.adapter.ValueObjectFactory;
 import org.synchronoss.cpo.jdbc.meta.JdbcCpoMetaDescriptor;
 import org.synchronoss.cpo.meta.domain.CpoClass;
 import org.testng.annotations.AfterClass;
