+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/jta/JdbcXaResourceTest.java
@@ -36,9 +36,9 @@ import org.synchronoss.cpo.CpoAdapter;
 import org.synchronoss.cpo.CpoAdapterFactoryManager;
 import org.synchronoss.cpo.CpoOrderBy;
 import org.synchronoss.cpo.helper.ExceptionHelper;
-import org.synchronoss.cpo.jdbc.JdbcStatics;
 import org.synchronoss.cpo.jdbc.ValueObject;
-import org.synchronoss.cpo.jdbc.ValueObjectFactory;
+import org.synchronoss.cpo.jdbc.adapter.JdbcStatics;
+import org.synchronoss.cpo.jdbc.adapter.ValueObjectFactory;
 import org.synchronoss.cpo.jta.CpoXaError;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
